Video_models/Crowd_analysis_models/CrowdMAC/requirements.txt
einops
hydra-core
joblib
matplotlib
numpy
omegaconf
opencv-python
pandas
pillow
scikit-learn
scikit-image
scipy
tensorboard
timm
tqdm
wandb
easydict